backend/requirements.txt
fastapi
uvicorn[standard]
pyswisseph==2.10.03
pandas
